% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpsg_getsolution.R
\name{rpsg_getsolution_one_problem}
\alias{rpsg_getsolution_one_problem}
\title{Convert character PSG solution report into list witn numeric data in the case of one
PSG prblem (only one output report).}
\usage{
rpsg_getsolution_one_problem(problem.res)
}
\arguments{
\item{problem.res}{list with solution results of the optimization problem obtained by \code{\link{rpsg_solver}}}
}
\value{
output.list list with solution results
}
\description{
Convert character PSG solution report into list witn numeric data in the case of one
PSG prblem (only one output report).
}
\keyword{internal}

